modification of group member

+++ app/Http/Controllers/Admin/School/GroupmemberController.php
@@ -1,17 +1,19 @@
 <?php
 
-namespace App\Http\Controllers\Admin;
+namespace App\Http\Controllers\Admin\School;
 
 use App\Model\School\EducationInstitute as EduInstitute; 
 use App\Http\Controllers\Admin\BaseController;
 use App\CompanyUser;
 use App\Http\Requests\Admin\GroupMemberRequest;
 use App\Model\School\Groupmember;
+use Carbon\Carbon;
+use Validator; 
 use Illuminate\Http\Request;
 
 class GroupmemberController extends BaseController
 {
-    public function getMember($tour_code,$user_type, $tour_type){
+    public function getMember($tour_type, $tour_code,$user_type){
         try{
             $where = ['tour_id'=>$tour_code,'user_type'=>$user_type, 'tour_type'=>$tour_type];
             $groupmember = Groupmember::where($where)->get();
@@ -26,7 +28,7 @@ class GroupmemberController extends BaseController
             return $this->sendError($e->getMessage(), 500);
         }
     }
-    public function getMemberPending($tour_code,$user_type, $tour_type, $pending){
+    public function getMemberPending($tour_type, $tour_code,$user_type, $pending){
         try{
             $where = ['tour_id'=>$tour_code,'user_type'=>$user_type,'payment_status'=>$pending, 'tour_type'=>$tour_type];
             $groupmember = Groupmember::where($where)->get();
@@ -42,8 +44,16 @@ class GroupmemberController extends BaseController
             return $this->sendError($e->getMessage(), 500);
         }
     }
-    public function updateMember(Request $request){
+    public function updateMember($tour_type, Request $request){
         try{
+            $validator = Validator::make($request->all(), [ 
+                'id'=>'required|exists:groupmembers,id', 
+            ]);
+    
+            if ($validator->fails()) { 
+                return response()->json(['error'=>$validator->errors()], 401);            
+            }
+
             $group_member = Groupmember::where('id',$request->id)->firstOrFail();
             $group_member->first_name = $request->first_name??$group_member->first_name;
             $group_member->last_name = $request->last_name??$group_member->last_name;
@@ -61,15 +71,22 @@ class GroupmemberController extends BaseController
             return $this->sendError($e->getMessage(), 500);
         }
     }
-    public function destroyMember(Request $request){
+    public function destroyMember($tour_type, Request $request){
+        $validator = Validator::make($request->all(), [ 
+            'id'=>'required|exists:groupmembers,id', 
+        ]);
+
+        if ($validator->fails()) { 
+            return response()->json(['error'=>$validator->errors()], 401);            
+        }
         $groupmember = Groupmember::where('id',$request->id)->firstOrFail();
         $groupmember->delete();
-        return response()->json('successfully delete');
+        return response()->json('successfully deleted');
     }
-    public function addMember(GroupMemberRequest $request){
+    public function addMember($tour_type, GroupMemberRequest $request){
         try{
             $data = array();
-            if($request->tour_type == "school"){
+            if($tour_type == "school"){
                 $edu_institute = EduInstitute::where('school_id', $request->school_id??NULL)->first();
                 if(!$edu_institute){
                     return $this->sendError("Invalid user", 404);
@@ -77,7 +94,7 @@ class GroupmemberController extends BaseController
                 $data['edu_institute_id'] = $edu_institute->id??NULL;
                 $data['school_id'] = $request->school_id??NULL;
             }
-            else if($request->tour_type == "corporate"){
+            else if($tour_type == "corporate"){
                 $company_user = CompanyUser::where('company_id', $request->company_id??NULL)->first();
                 if(!$company_user){
                     return $this->sendError("Invalid user", 404);
@@ -89,7 +106,7 @@ class GroupmemberController extends BaseController
                 $data['family_user_id'] = $request->family_user_id??NULL;
             }     
             $data['user_type'] = $request->user_type??NULL;
-            $data['tour_type'] = $request->tour_type??NULL;
+            $data['tour_type'] = $tour_type;
             $data['tour_id'] = $request->tour_id??'';
             if($request->details){
                 foreach ($request->details as $detail) {
